+++ src/applications/edu-benefits/1919/containers/PrivacyPolicy.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { querySelectorWithShadowRoot } from 'platform/utilities/ui/webComponents';
 import { VaModal } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
 import { useSelector } from 'react-redux';
-import PrivacyActStatement from './PrivacyActStatement';
+import PrivacyActStatement from '../components/PrivacyActStatement';
 import { getTitle } from '../helpers';
 
 const PrivacyPolicy = () => {
